Rename test view provider class/interface
