+++ src/common/configs/assessment-visualization-configuration.ts
@@ -3,7 +3,7 @@
 import { ToggleActionPayload } from 'background/actions/action-payloads';
 import { UniquelyIdentifiableInstances } from 'background/instance-identifier-generator';
 import { ScanIncompleteWarningId } from 'common/types/store-data/scan-incomplete-warnings';
-import { ITestViewContainerProvider } from 'DetailsView/components/itest-view-container-provider';
+import { TestViewContainerProvider } from 'DetailsView/components/test-view-container-provider';
 import { Analyzer, AnalyzerConfiguration } from '../../injected/analyzers/analyzer';
 import { AnalyzerProvider } from '../../injected/analyzers/analyzer-provider';
 import { VisualizationInstanceProcessorCallback } from '../../injected/visualization-instance-processor';
@@ -46,5 +46,5 @@ export interface AssessmentVisualizationConfiguration {
     getInstanceIdentiferGenerator: (
         testStep?: string,
     ) => (instance: UniquelyIdentifiableInstances) => string;
-    getTestViewContainer: (provider: ITestViewContainerProvider, props: any) => JSX.Element;
+    getTestViewContainer: (provider: TestViewContainerProvider, props: any) => JSX.Element;
 }
