+++ tsconfig.strictNullChecks.json
@@ -19,6 +19,7 @@
         "src/DetailsView/components/assessment-view-update-handler.ts",
         "src/DetailsView/components/automated-checks-command-bar.tsx",
         "src/DetailsView/components/command-bar-buttons-menu.tsx",
+        "src/DetailsView/components/default-test-view-container-provider.tsx",
         "src/DetailsView/components/details-list-issues-view.tsx",
         "src/DetailsView/components/details-view-command-bar.tsx",
         "src/DetailsView/components/details-view-content-with-local-state.tsx",
@@ -64,7 +65,6 @@
         "src/DetailsView/components/status-icon.tsx",
         "src/DetailsView/components/tab-stops-failed-instance-section.tsx",
         "src/DetailsView/components/tab-stops/tab-stops-instance-section-props-factory.tsx",
-        "src/DetailsView/components/test-view-container-provider.tsx",
         "src/DetailsView/components/test-view-container.tsx",
         "src/DetailsView/components/test-view.ts",
         "src/DetailsView/details-view-body.tsx",
